requirements.txt
aiogram
psycopg2-binary
environs
openpyxl
asyncio
pymysql